backend/pyproject.toml
[tool.poetry]
name = "xChat"
version = "0.1.0"
description = ""
authors = ["Carlo Stanciu <carlo@ezalabs.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
openai = "^0.28.0"
fastapi = "^0.103.1"
pydantic = "^2.3.0"
langchain = "^0.0.310"
uvicorn = "^0.23.2"
beautifulsoup4 = "^4.12.2"
tiktoken = "^0.4.0"
psycopg2 = "^2.9.7"
lxml = "^4.9.3"
supabase = "1.2.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
